Updated to resolve current enemy targeted
If anyone ever reads this, know that I am so incredibly tired. The hours it took to get this to work were outstanding. IF you would like to know, elden ring has no stable address for target data. I had to make a cheat engine script that finds camera data, and make space to store it. Sounds simple but in practice and to get it in a way it can run headless was hell on earth, but it was so worth it because it works and I love it.